+++ src/core/execution/SAMMissileExecution.ts
@@ -1,3 +1,4 @@
+import { NukeType } from "../AnalyticsSchemas";
 import {
   Execution,
   Game,
@@ -9,7 +10,6 @@ import {
 import { TileRef } from "../game/GameMap";
 import { AirPathFinder } from "../pathfinding/PathFinding";
 import { PseudoRandom } from "../PseudoRandom";
-import { NukeType } from "../Schemas";
 
 export class SAMMissileExecution implements Execution {
   private active = true;
